requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
Django==4.0.6
django-cors-headers==3.13.0
django-crontab==0.7.1
django-extensions==3.2.0
django-filter==22.1
django-nose==1.4.7
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.7.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
nose==1.3.7
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
tomli==2.0.1
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
